Move to a database-centric model
This makes it way way easier to jump around without all the crazy hoops that used to be required in only the file-based structure.
Keeping the file structure is good for validating the metadata YAML files, and we can move all result-centric stuff into the db model classes.

TODO: Put in a new datacheck model